ai-service/requirements.txt
fastapi
uvicorn
langchain-core
langchain-groq
langgraph
supabase
storage3
pydantic
python-dotenv
httpx
pytest
gunicorn
uvicorn-worker